🚚 chore: Move os DTOS para um novo diretório 'request'
