+++ src/main/java/br/com/danllopes/usermanagment/dtos/request/AuthenticationDTO.java
@@ -1,4 +1,4 @@
-package br.com.danllopes.usermanagment.dtos;
+package br.com.danllopes.usermanagment.dtos.request;
 
 import jakarta.validation.constraints.NotBlank;
